requirements.txt
numpy
pydub
ffmpeg-python
matplotlib
